requirements.txt
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
mss==7.0.1
nltk==3.8.1
numpy==1.24.1
opencv-python==4.7.0.68
packaging==22.0
Pillow==9.4.0
pyparsing==3.0.9
pyserial==3.5
pytesseract==0.3.10
python-dateutil==2.8.2
regex==2022.10.31
six==1.16.0
tqdm==4.64.1
